Redesigned process, adapted testdriver

+++ src/main/java/tech/stackable/t2/cluster/ClusterService.java
@@ -3,13 +3,12 @@ package tech.stackable.t2.cluster;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
-import java.time.Duration;
-import java.time.LocalDateTime;
-import java.util.Collection;
+import java.text.MessageFormat;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -17,13 +16,14 @@ import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Repository;
+import org.springframework.util.CollectionUtils;
 
-import io.micrometer.core.instrument.Counter;
 import tech.stackable.t2.ansible.AnsibleResult;
 import tech.stackable.t2.ansible.AnsibleService;
+import tech.stackable.t2.api.ClusterNotFoundException;
+import tech.stackable.t2.api.ClusterNotRunningException;
+import tech.stackable.t2.api.IllegalClusterStateTransitionException;
 import tech.stackable.t2.domain.Cluster;
 import tech.stackable.t2.domain.Status;
 import tech.stackable.t2.files.FileService;
@@ -38,8 +38,6 @@ public class ClusterService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ClusterService.class);
 
-    private static final Duration CLEANUP_INACTIVITY_THRESHOLD = Duration.ofDays(1);
-
     @Autowired
     private FileService fileService;
 
@@ -48,18 +46,6 @@ public class ClusterService {
 
     @Autowired
     private AnsibleService ansibleService;
-
-    @Autowired
-    @Qualifier("clustersRequestedCounter")
-    private Counter clustersRequestedCounter;
-
-    @Autowired
-    @Qualifier("clustersCreatedCounter")
-    private Counter clustersCreatedCounter;
-
-    @Autowired
-    @Qualifier("clustersTerminatedCounter")
-    private Counter clustersTerminatedCounter;
 
     /**
      * This is the main storage for all clusters (by UUID).
@@ -73,8 +59,20 @@ public class ClusterService {
      * 
      * @return list of all clusters
      */
-    public Collection<Cluster> getClusters() {
-        return this.clusters.values();
+    public List<Cluster> getClusters() {
+        return this.getClusters(null);
+    }
+
+    /**
+     * Get list of clusters filtered by status.
+     * 
+     * @param statusFilter Set of Status values to be included in the list.
+     * @return list of clusters, filtered by status
+     */
+    public List<Cluster> getClusters(Set<Status> statusFilter) {
+        return this.clusters.values().stream()
+                .filter(cluster -> (CollectionUtils.isEmpty(statusFilter) || statusFilter.contains(cluster.getStatus())))
+                .collect(Collectors.toList());
     }
 
     /**
@@ -88,6 +86,38 @@ public class ClusterService {
     }
 
     /**
+     * Set the status of a cluster.
+     * 
+     * There are only two allowed status changes:
+     * 
+     * <ul>
+     * <li>{@link Status#LAUNCH_FAILED} to {@link Status#TERMINATED_MANUALLY}</li>
+     * <li>{@link Status#TERMINATION_FAILED} to {@link Status#TERMINATED_MANUALLY}</li>
+     * </ul>
+     * 
+     * @param id     ID of the cluster
+     * @param status status to set on the cluster
+     * @return cluster with the given ID and new status
+     */
+    public Cluster setClusterStatus(UUID id, Status status) {
+        synchronized (this.clusters) {
+            Cluster cluster = this.clusters.get(id);
+            if (cluster == null) {
+                throw new ClusterNotFoundException(MessageFormat.format("No cluster found with ID {0}", id));
+            }
+            synchronized (cluster) {
+                if ((cluster.getStatus() == Status.LAUNCH_FAILED || cluster.getStatus() == Status.TERMINATION_FAILED) && status == Status.TERMINATED_MANUALLY) {
+                    cluster.setStatus(status);
+                    cluster.addEvent(MessageFormat.format("Status set to {0}", status));
+                    return cluster;
+                } else {
+                    throw new IllegalClusterStateTransitionException(MessageFormat.format("The status of the cluster {0} cannot be set to {1}.", id, status));
+                }
+            }
+        }
+    }
+
+    /**
      * Start the creation of a new cluster.
      * 
      * @param clusterDefinition definition of the new cluster
@@ -97,66 +127,109 @@ public class ClusterService {
         synchronized (this.clusters) {
 
             Cluster cluster = new Cluster();
-            cluster.setStatus(Status.CREATION_STARTED);
-            this.clustersRequestedCounter.increment();
 
-            Path workingDirectory = this.fileService.workingDirectory(cluster.getId());
+            synchronized (cluster) {
+                Path workingDirectory = this.fileService.workingDirectory(cluster.getId());
+                cluster.addEvent(MessageFormat.format("Creating working directory {0}...", workingDirectory));
+                this.fileService.createWorkingDirectory(workingDirectory, clusterDefinition);
+                cluster.addEvent(MessageFormat.format("Created working directory {0}.", workingDirectory));
+                
+                cluster.setStatus(Status.LAUNCHING);
 
-            this.fileService.createWorkingDirectory(workingDirectory, clusterDefinition);
-            cluster.setStatus(Status.WORKING_DIR_CREATED);
-            clusters.put(cluster.getId(), cluster);
+                clusters.put(cluster.getId(), cluster);
 
-            // This thread must be started if the cluster launch fails after terraform apply
-            // has started as there may be created resources
-            // which have to be torn down...
-            Thread tearDownOnFailure = new Thread(() -> {
-                this.terraformService.destroy(cluster.getId());
-                this.fileService.cleanUpWorkingDirectory(workingDirectory);
-            });
+                // Launching cluster in new thread
+                new Thread(() -> {
 
-            new Thread(() -> {
+                    TerraformResult terraformResult = null;
 
-                TerraformResult terraformResult = null;
+                    cluster.addEvent("Terraform init started.");
+                    terraformResult = this.terraformService.init(cluster.getId());
+                    if (terraformResult == TerraformResult.ERROR) {
+                        cluster.addEvent(MessageFormat.format("Terraform init failed with result {0}", terraformResult));
+                        cluster.addEvent("Working directory cleanup started...");
+                        this.fileService.cleanUpWorkingDirectory(workingDirectory);
+                        cluster.addEvent("Working directory cleaned up.");
+                        cluster.setStatus(Status.LAUNCH_FAILED);
+                        return;
+                    }
+                    cluster.addEvent("Terraform init successful.");
 
-                cluster.setStatus(Status.TERRAFORM_INIT);
-                terraformResult = this.terraformService.init(cluster.getId());
-                if (terraformResult == TerraformResult.ERROR) {
-                    cluster.setStatus(Status.TERRAFORM_INIT_FAILED);
-                    this.fileService.cleanUpWorkingDirectory(workingDirectory);
-                    return;
-                }
+                    cluster.addEvent("Terraform plan started.");
+                    terraformResult = this.terraformService.plan(cluster.getId());
+                    if (terraformResult == TerraformResult.ERROR) {
+                        cluster.addEvent(MessageFormat.format("Terraform plan failed with result {0}", terraformResult));
+                        cluster.addEvent("Working directory cleanup started...");
+                        this.fileService.cleanUpWorkingDirectory(workingDirectory);
+                        cluster.addEvent("Working directory cleaned up.");
+                        cluster.setStatus(Status.LAUNCH_FAILED);
+                        return;
+                    }
+                    cluster.addEvent("Terraform plan successful.");
 
-                cluster.setStatus(Status.TERRAFORM_PLAN);
-                terraformResult = this.terraformService.plan(cluster.getId());
-                if (terraformResult == TerraformResult.ERROR) {
-                    cluster.setStatus(Status.TERRAFORM_PLAN_FAILED);
-                    this.fileService.cleanUpWorkingDirectory(workingDirectory);
-                    return;
-                }
+                    cluster.addEvent("Terraform apply started.");
+                    terraformResult = this.terraformService.apply(cluster.getId());
+                    if (terraformResult == TerraformResult.ERROR) {
+                        cluster.addEvent(MessageFormat.format("Terraform apply failed with result {0}", terraformResult));
+                        cleanupAfterFailedLaunch(cluster);
+                        return;
+                    }
+                    cluster.addEvent("Terraform apply successful.");
 
-                cluster.setStatus(Status.TERRAFORM_APPLY);
-                terraformResult = this.terraformService.apply(cluster.getId());
-                if (terraformResult == TerraformResult.ERROR) {
-                    cluster.setStatus(Status.TERRAFORM_APPLY_FAILED);
-                    tearDownOnFailure.start();
-                    return;
-                }
+                    cluster.addEvent("Ansible launch started.");
+                    AnsibleResult ansibleResult = this.ansibleService.launch(cluster.getId());
+                    if (ansibleResult == AnsibleResult.ERROR) {
+                        cluster.addEvent("Ansible launch failed.");
+                        cleanupAfterFailedLaunch(cluster);
+                        return;
+                    }
+                    cluster.addEvent("Ansible launch successful.");
 
-                cluster.setStatus(Status.ANSIBLE_PROVISIONING);
-                AnsibleResult ansibleResult = this.ansibleService.run(cluster.getId());
-                if (ansibleResult == AnsibleResult.ERROR) {
-                    cluster.setStatus(Status.ANSIBLE_FAILED);
-                    tearDownOnFailure.start();
-                    return;
-                }
+                    cluster.addEvent("Cluster up and running!");
+                    cluster.setStatus(Status.RUNNING);
 
-                cluster.setStatus(Status.RUNNING);
-                this.clustersCreatedCounter.increment();
-
-            }).start();
+                }).start();
+            }
 
             return cluster;
         }
+    }
+
+    /**
+     * Cleans up after a failed cluster launch and eventually sets the status to {@link Status#LAUNCH_FAILED}.
+     * 
+     * The cleanup happens in a separate thread!
+     * 
+     * @param cluster cluster to be cleaned up.
+     */
+    private void cleanupAfterFailedLaunch(Cluster cluster) {
+        new Thread(() -> {
+            synchronized (cluster) {
+                Path workingDirectory = this.fileService.workingDirectory(cluster.getId());
+
+                cluster.addEvent("Ansible cleanup started.");
+                AnsibleResult ansibleResult = this.ansibleService.cleanup(cluster.getId());
+                if (ansibleResult == AnsibleResult.SUCCESS) {
+                    cluster.addEvent("Ansible cleanup successful.");
+                } else {
+                    cluster.addEvent("Ansible cleanup failed.");
+                }
+
+                cluster.addEvent("Terraform destroy started.");
+                TerraformResult terraformResult = this.terraformService.destroy(cluster.getId());
+                if (terraformResult == TerraformResult.SUCCESS) {
+                    cluster.addEvent("Terraform destroy successful.");
+                } else {
+                    cluster.addEvent(MessageFormat.format("Terraform destroy failed with result {0}", terraformResult));
+                }
+
+                cluster.addEvent("Working directory cleanup started...");
+                this.fileService.cleanUpWorkingDirectory(workingDirectory);
+                cluster.addEvent("Working directory cleaned up.");
+
+                cluster.setStatus(Status.LAUNCH_FAILED);
+            }
+        }).start();
     }
 
     /**
@@ -171,22 +244,47 @@ public class ClusterService {
             if (cluster == null) {
                 return Optional.empty();
             }
-            cluster.setStatus(Status.DELETION_STARTED);
 
-            new Thread(() -> {
-
-                Path workingDirectory = this.fileService.workingDirectory(cluster.getId());
-
-                cluster.setStatus(Status.TERRAFORM_DESTROY);
-                TerraformResult terraformResult = this.terraformService.destroy(cluster.getId());
-                if (terraformResult == TerraformResult.ERROR) {
-                    cluster.setStatus(Status.TERRAFORM_DESTROY_FAILED);
-                    return;
+            synchronized (cluster) {
+                if (cluster.getStatus() != Status.RUNNING) {
+                    throw new ClusterNotRunningException(MessageFormat.format("The cluster {0} is not running.", id));
                 }
-                this.fileService.cleanUpWorkingDirectory(workingDirectory);
-                cluster.setStatus(Status.TERMINATED);
-                this.clustersTerminatedCounter.increment();
-            }).start();
+
+                cluster.setStatus(Status.TERMINATING);
+
+                new Thread(() -> {
+                    synchronized (cluster) {
+                        Path workingDirectory = this.fileService.workingDirectory(cluster.getId());
+
+                        cluster.addEvent("Ansible cleanup started.");
+                        AnsibleResult ansibleResult = this.ansibleService.cleanup(cluster.getId());
+                        if (ansibleResult == AnsibleResult.SUCCESS) {
+                            cluster.addEvent("Ansible cleanup successful.");
+                        } else {
+                            cluster.addEvent("Ansible cleanup failed.");
+                        }
+
+                        cluster.addEvent("Terraform destroy started.");
+                        TerraformResult terraformResult = this.terraformService.destroy(cluster.getId());
+                        if (terraformResult == TerraformResult.SUCCESS) {
+                            cluster.addEvent("Terraform destroy successful.");
+                        } else {
+                            cluster.addEvent(MessageFormat.format("Terraform destroy failed with result {0}", terraformResult));
+                        }
+
+                        cluster.addEvent("Working directory cleanup started...");
+                        this.fileService.cleanUpWorkingDirectory(workingDirectory);
+                        cluster.addEvent("Working directory cleaned up.");
+
+                        if (ansibleResult != AnsibleResult.SUCCESS || terraformResult != TerraformResult.SUCCESS) {
+                            cluster.setStatus(Status.TERMINATION_FAILED);
+                            return;
+                        }
+
+                        cluster.setStatus(Status.TERMINATED);
+                    }
+                }).start();
+            }
 
             return Optional.of(cluster);
         }
@@ -266,44 +364,44 @@ public class ClusterService {
 //        return bos.toByteArray();
 //    }
 
-    /**
-     * Cleans up list of clusters on every hour.
-     * 
-     * This method removes all clusters from the in-memory data which are {@link #readyForCleanup(Cluster)}
-     */
-    @Scheduled(cron = "0 0 * * * *")
-    private void cleanup() {
-        LOGGER.info("cleaning up clusters ...");
-        List<UUID> clustersToDelete = this.clusters.values()
-                .stream()
-                .filter(ClusterService::readyForCleanup)
-                .map(Cluster::getId)
-                .collect(Collectors.toList());
-
-        LOGGER.info("Found {} clusters to be cleaned up.", clustersToDelete.size());
-
-        synchronized (clusters) {
-            clustersToDelete.forEach(id -> {
-                LOGGER.info("Cluster {} will be cleaned up.", id);
-                this.clusters.remove(id);
-            });
-        }
-
-        LOGGER.info("cleaned up {} clusters.", clustersToDelete.size());
-    }
-
-    /**
-     * Decides if a given Cluster is ready to be cleaned up.
-     * 
-     * We assume that clusters that are not in state {@link Status#RUNNING} and haven't changed their state for a day are
-     * ready to be removed.
-     * 
-     * @param cluster cluster to check
-     * @return Is the given cluster ready to be cleaned up?
-     */
-    private static boolean readyForCleanup(Cluster cluster) {
-        return !(cluster.getStatus() == Status.RUNNING)
-                && Duration.between(cluster.getLastChangedAt(), LocalDateTime.now())
-                        .compareTo(CLEANUP_INACTIVITY_THRESHOLD) > 0;
-    }
+//    /**
+//     * Cleans up list of clusters on every hour.
+//     * 
+//     * This method removes all clusters from the in-memory data which are {@link #readyForCleanup(Cluster)}
+//     */
+//    @Scheduled(cron = "0 0 * * * *")
+//    private void cleanup() {
+//        LOGGER.info("cleaning up clusters ...");
+//        List<UUID> clustersToDelete = this.clusters.values()
+//                .stream()
+//                .filter(ClusterService::readyForCleanup)
+//                .map(Cluster::getId)
+//                .collect(Collectors.toList());
+//
+//        LOGGER.info("Found {} clusters to be cleaned up.", clustersToDelete.size());
+//
+//        synchronized (clusters) {
+//            clustersToDelete.forEach(id -> {
+//                LOGGER.info("Cluster {} will be cleaned up.", id);
+//                this.clusters.remove(id);
+//            });
+//        }
+//
+//        LOGGER.info("cleaned up {} clusters.", clustersToDelete.size());
+//    }
+//
+//    /**
+//     * Decides if a given Cluster is ready to be cleaned up.
+//     * 
+//     * We assume that clusters that are not in state {@link Status#RUNNING} and haven't changed their state for a day are
+//     * ready to be removed.
+//     * 
+//     * @param cluster cluster to check
+//     * @return Is the given cluster ready to be cleaned up?
+//     */
+//    private static boolean readyForCleanup(Cluster cluster) {
+//        return !(cluster.getStatus() == Status.RUNNING)
+//                && Duration.between(cluster.getLastChangedAt(), LocalDateTime.now())
+//                        .compareTo(CLEANUP_INACTIVITY_THRESHOLD) > 0;
+//    }
 }
